qa_pipeline_service/requirements.txt
fastapi
uvicorn
pdfminer.six
faiss-cpu
langchain-community
python-multipart
pydantic
fastembed
sentence-transformers
langchain-huggingface